pyproject.toml
[project]
name = "af3builder"
version = "0.1.0"
authors = [{name = "Ronald Oellers", email = "oellers@ie-freiburg.mpg.de"}]
description = "AlphaFold3 input preparation toolkit"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.28",
    "biopython>=1.81",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "click>=8.0"
]

[project.scripts]
af3build = "af3builder.cli:main"
